+++ example/.dart_tool/flutter_build/a0e32b7d00e7a6c10a8dc63634a3fd1f/flutter_assets.d
@@ -1,0 +1,1 @@
+ build/flutter_assets/fonts/MaterialIcons-Regular.ttf build/flutter_assets/AssetManifest.json build/flutter_assets/FontManifest.json build/flutter_assets/LICENSE:  /Users/zfrontier/Dev/zfrontier/flutter_webview_plugin/example/pubspec.yaml /usr/local/flutter/bin/cache/artifacts/material_fonts/MaterialIcons-Regular.ttf