+++ lib/flutter_webview_plugin.dart
@@ -1,6 +1,7 @@
 library flutter_webview_plugin;
 
 export 'src/base.dart';
-export 'src/js_channel.dart';
+export 'src/javascript_channel.dart';
+export 'src/javascript_message.dart';
 export 'src/navigation.dart';
 export 'src/webview_scaffold.dart';